+++ src/Command/NewsTagsMigrationCommand.php
@@ -10,9 +10,11 @@ namespace HeimrichHannot\MigrationBundle\Command;
 
 use Contao\CoreBundle\Command\AbstractLockedCommand;
 use Doctrine\DBAL\Query\QueryBuilder;
+use HeimrichHannot\UtilsBundle\Database\DatabaseUtil;
 use HeimrichHannot\UtilsBundle\Driver\DC_Table_Utils;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Symfony\Component\DependencyInjection\ContainerInterface;
@@ -35,20 +37,19 @@ class NewsTagsMigrationCommand extends AbstractLockedCommand
     protected $output;
 
     /**
-     * tl_news category field name.
-     *
-     * @var string
-     */
-    protected $field;
-    /**
      * @var ContainerInterface
      */
     private $container;
+    /**
+     * @var DatabaseUtil
+     */
+    private $databaseUtil;
 
-    public function __construct(ContainerInterface $container)
+    public function __construct(ContainerInterface $container, DatabaseUtil $databaseUtil)
     {
         parent::__construct();
-        $this->container = $container;
+        $this->container    = $container;
+        $this->databaseUtil = $databaseUtil;
     }
 
     /**
@@ -60,7 +61,7 @@ class NewsTagsMigrationCommand extends AbstractLockedCommand
             'Migration of database entries from tags module to codefog/tags-bundle.'
         );
 
-        $this->addArgument('field', InputArgument::OPTIONAL, 'What is the name of the tags field in tl_news (default: tags)?', 'tags');
+        $this->addOption('news-archive-ids', null, InputOption::VALUE_OPTIONAL, 'Restrict the command to news of certain archives; pass in a comma separated list of IDs (default: no restriction)?', '');
 
         parent::configure();
     }
@@ -72,10 +73,11 @@ class NewsTagsMigrationCommand extends AbstractLockedCommand
     {
         $this->container->get('contao.framework')->initialize();
 
-        $this->field = $input->getArgument('field');
-        $this->io = new SymfonyStyle($input, $output);
-        $this->input = $input;
-        $this->output = $output;
+        $this->newsArchiveIds = explode(',', $input->getOption('news-archive-ids'));
+
+        $this->io      = new SymfonyStyle($input, $output);
+        $this->input   = $input;
+        $this->output  = $output;
         $this->rootDir = $this->getContainer()->getParameter('kernel.project_dir');
 
         $this->queryBuilder = new QueryBuilder($this->getContainer()->get('doctrine.dbal.default_connection'));
@@ -90,13 +92,29 @@ class NewsTagsMigrationCommand extends AbstractLockedCommand
     {
         $this->queryBuilder->setParameter('from_table', 'tl_news');
 
-        $newsTags = $this->queryBuilder->select('*')->from('tl_tag')->where($this->queryBuilder->expr()->eq('tl_tag.from_table', ':from_table'))->execute()->fetchAll();
+        $this->queryBuilder->select('*')->from('tl_tag')->where($this->queryBuilder->expr()->eq('tl_tag.from_table', ':from_table'));
+
+        // restrict news archive ids
+        if (!empty($this->newsArchiveIds)) {
+            // retrieve news ids from the archives
+            $newsIds = [];
+
+            if (null !== ($news = $this->databaseUtil->findResultsBy('tl_news', ['tl_news.pid IN (' . implode(',', $this->newsArchiveIds) . ')'], [])) &&
+                $news->numRows > 0) {
+                $newsIds = $news->fetchEach('id');
+            }
+
+            // not in the condition above because if the condition is not set, all news_ids would be migrated!
+            $this->queryBuilder->andWhere($this->queryBuilder->expr()->in('tid', $newsIds));
+        }
+
+        $newsTags = $this->queryBuilder->execute()->fetchAll();
 
         if (false === $newsTags) {
             return 0;
         }
 
-        $tstamp = time();
+        $tstamp  = time();
         $newsIds = [];
 
         foreach ($newsTags as $newsTag) {
@@ -104,36 +122,36 @@ class NewsTagsMigrationCommand extends AbstractLockedCommand
 
             // 1st: create the tag if it not already exists
             if (false === $tag) {
-                $this->queryBuilder->setParameters(['tstamp' => $tstamp, 'name' => $newsTag['tag'], 'source' => 'app.news']);
+                $this->queryBuilder->setParameters(['tstamp' => $tstamp, 'name' => $newsTag['tag'], 'source' => 'news_tag_manager']);
                 $this->queryBuilder->resetQueryParts();
                 $this->queryBuilder->insert('tl_cfg_tag')->values(['tstamp' => ':tstamp', 'name' => ':name', 'source' => ':source'])->execute();
             }
 
             if (false === ($tag = $this->findTagByName($newsTag['tag']))) {
-                $this->output->writeln('<error>Error: Could not create news tag with name: "'.$newsTag['tag'].'"</error>');
+                $this->output->writeln('<error>Error: Could not create news tag with name: "' . $newsTag['tag'] . '"</error>');
 
                 continue;
             }
 
-            // 2nd: create an unique alias
+            // 2nd: create a unique alias
             if (empty($tag['alias'])) {
-                $dc = new DC_Table_Utils('tl_cfg_tag');
-                $dc->activeRecord = (object) $tag;
+                $dc               = new DC_Table_Utils('tl_cfg_tag');
+                $dc->activeRecord = (object)$tag;
 
-                $this->queryBuilder->setParameters(['alias' => $this->container->get('codefog_tags.listener.data_container.tag')->generateAlias('', $dc), 'id' => $tag['id']]);
+                $this->queryBuilder->setParameters(['alias' => $this->container->get('codefog_tags.listener.data_container.tag')->onAliasSaveCallback('', $dc), 'id' => $tag['id']]);
                 $this->queryBuilder->update('tl_cfg_tag')->set('alias', ':alias')->where($this->queryBuilder->expr()->eq('id', ':id'))->execute();
             }
 
             // 3rd: delete/cleanup news / news tag relations from previous imports
             $this->queryBuilder->resetQueryParts();
             $this->queryBuilder->setParameters(['news_id' => $newsTag['tid'], 'tag_id' => $tag['id']]);
-            $this->queryBuilder->delete('tl_news_tags')->where($this->queryBuilder->expr()->eq('news_id', ':news_id'))->andWhere($this->queryBuilder->expr()->eq('cfg_tag_id', ':tag_id'))->execute();
+            $this->queryBuilder->delete('tl_cfg_tag_news')->where($this->queryBuilder->expr()->eq('news_id', ':news_id'))->andWhere($this->queryBuilder->expr()->eq('cfg_tag_id', ':tag_id'))->execute();
 
             // 4th: create news / news tag relations
             $this->queryBuilder->resetQueryParts();
             $this->queryBuilder->setParameters(['news_id' => $newsTag['tid'], 'tag_id' => $tag['id']]);
 
-            if ($this->queryBuilder->insert('tl_news_tags')->values(['news_id' => ':news_id', 'cfg_tag_id' => ':tag_id'])->execute()) {
+            if ($this->queryBuilder->insert('tl_cfg_tag_news')->values(['news_id' => ':news_id', 'cfg_tag_id' => ':tag_id'])->execute()) {
                 $newsIds[$newsTag['tid']][] = $tag['id'];
             }
         }
@@ -142,7 +160,7 @@ class NewsTagsMigrationCommand extends AbstractLockedCommand
             $this->queryBuilder->setParameters(['id' => $id, 'tags' => serialize($tags)]);
 
             if ($this->queryBuilder->update('tl_news')->set('tags', ':tags')->where($this->queryBuilder->expr()->eq('id', ':id'))->execute()) {
-                $this->output->writeln('<info>Successfully migrated tags for news with id: "'.$id.'"</info>');
+                $this->output->writeln('<info>Successfully migrated tags for news with id: "' . $id . '"</info>');
             }
         }
 
@@ -159,7 +177,7 @@ class NewsTagsMigrationCommand extends AbstractLockedCommand
     protected function findTagByName($name)
     {
         $this->queryBuilder->resetQueryParts();
-        $this->queryBuilder->setParameters(['name' => $name, 'source' => 'app.news']);
+        $this->queryBuilder->setParameters(['name' => $name, 'source' => 'news_tag_manager']);
 
         return $this->queryBuilder->select('*')->from('tl_cfg_tag')->where($this->queryBuilder->expr()->eq('name', ':name'))->andWhere($this->queryBuilder->expr()->eq('source', ':source'))->setMaxResults(1)->execute()->fetch();
     }